You have no idea.
The fewest docs possible.

Anyway, now we have a method for sending packets if we need that (heh, waste of time) and the power is sent to the client! Quel bonheur!

Also fixed a mini mistake and added a setEnergy function for the EnergyHandler

If you need to send things over packets, just call me!

Good night!