Really, what have I not done? Lets step through the changes.
First and foremost, the most important change that has been made is the COMPLETE AND UTTER overhaul of the genetics system, not incorporating a "left" and "right" side (pairs of chromosomes) and an "expressed" version which is the somewhat randomised mix of the "left" and "right" genes, but also, which is the other MAJOR change, dominance (recessive and dominant genes) which choose which gene gets expressed, and are passed down.

Haploid-esque gametes are generated when a baby is born from GM parents, somewhat realistically emulating real life, as well as the "left" and "right" versions of the chromosomes for the child _are_ copies of the gametes produced by the parents, which are then evaluated to create the "expressed" version of the genes. Lots of code has also been cut from BabySpawnEventHandler, unimportant stuff.

Of course, all of the required stuff has been updated to use this new system, such as ItemSyringe and BasePetriDish so that they store all the required info still. Also, although not completely, CommandSummonGeneticallyModified has been updated for this new system. MobSpawnEventHandler has been COMPLETELY redesigned, now calling a switch statement for specific genes, and with a standard operating procedure and arguments for every gene handler which are standardised. At some point, I might create a system which makes especially the "Attributes" NBTTagList traversing more efficient, but for now it works and does not produce noticeable lag. Of course, GeneticsStorage has been updated, with the help of a new set of methods in NBTParser, which convert between the NBTTagList of (mostly) 3 NBTTagLists of NBTTagCompounds and List<Map<String, Double>>.

There are a couple of new files, "GeneticsMismatchException.java", "EntityHandlerGeneric.java", and "GeneticsSide.java". The first is a standard bare RuntimeException class which gets thrown if (normally impossible) the Gene and Gene Dominance maps do not contain the same keys (as that would make either some genes not have a dominance, or some dominances would be "ethereal"). Speaking of which, why have 2 different maps for dominance and genes? Well, it's by far the easiest to implement in NBT and does not rely on that much more code. Also, many times you'd just be interested in the genes, so there's no reason to bundle dominance in with that. Truthfully, it was also the easiest to do, so why not? The second new class, "EntityHandlerGeneric.java", handles some (and at some point will handle all) the generic entity genes, such as Health and MovementSpeed. The final class is an Enum which holds the required "sides" for setGenes (when editing particular areas of genes, such as only the expressed version, as opposed to the setAllGenes which takes a list and edits them all).

I had to wrestle with some bugs, ending up with me adding the "hasBeenEdited" field, otherwise at every login, the same operation would be done on the entity (so doubling would have gotten out of hand very quickly).

Finally, but not least by any means, I took part in some cleaning up, so some files unrelated to what I have just outlined have been just very slightly edited.

I hope you enjoy!
Good Night!